Add short-tag alias for BC
